requirements.txt
fastapi
uvicorn
google-adk
python-dotenv
truststore